Remove guide and guidemodel classes
