+++ nebula/ui/components/VPNExpandableAppList.qml
@@ -8,6 +8,7 @@ import QtQuick.Controls 2.14
 import QtQuick.Layouts 1.14
 
 import Mozilla.VPN 1.0
+import Mozilla.VPN.qmlcomponents 1.0
 import components.forms 0.1
 
 ColumnLayout {
