+++ src/addons/addon.h
@@ -16,6 +16,7 @@ class Addon : public QObject {
 
   Q_PROPERTY(QString id READ id CONSTANT)
   Q_PROPERTY(QString name MEMBER m_name CONSTANT)
+  Q_PROPERTY(QString type READ type CONSTANT)
 
  public:
   static Addon* create(QObject* parent, const QString& manifestFileName);
@@ -25,17 +26,19 @@ class Addon : public QObject {
   ~Addon();
 
   const QString& id() const { return m_id; }
+  const QString& type() const { return m_type; }
 
   void retranslate();
 
  protected:
   Addon(QObject* parent, const QString& manifestFileName, const QString& id,
-        const QString& name);
+        const QString& name, const QString& type);
 
  private:
   const QString m_manifestFileName;
   const QString m_id;
   const QString m_name;
+  const QString m_type;
 
   QTranslator m_translator;
 };
