+++ src/addons/addonguide.cpp
@@ -5,7 +5,6 @@
 #include "addonguide.h"
 #include "leakdetector.h"
 #include "logger.h"
-#include "models/guidemodel.h"
 
 #include <QJsonObject>
 
@@ -17,22 +16,41 @@ Logger logger(LOG_MAIN, "AddonGuide");
 Addon* AddonGuide::create(QObject* parent, const QString& manifestFileName,
                           const QString& id, const QString& name,
                           const QJsonObject& obj) {
-  if (!GuideModel::instance()->createFromJson(id, obj["guide"].toObject())) {
-    logger.warning() << "Unable to add the guide";
+  QJsonObject guideObj = obj["guide"].toObject();
+
+  QString guideId = guideObj["id"].toString();
+  if (guideId.isEmpty()) {
+    logger.warning() << "Empty ID for guide";
     return nullptr;
   }
 
-  return new AddonGuide(parent, manifestFileName, id, name);
+  AddonGuide* guide = new AddonGuide(parent, manifestFileName, id, name);
+  auto guard = qScopeGuard([&] { guide->deleteLater(); });
+
+  guide->m_titleId = QString("guide.%1.title").arg(guideId);
+  guide->m_subtitleId = QString("guide.%1.subtitle").arg(guideId);
+
+  guide->m_image = guideObj["image"].toString();
+  if (guide->m_image.isEmpty()) {
+    logger.warning() << "Empty image for guide";
+    return nullptr;
+  }
+
+  guide->m_composer =
+      Composer::create(guide, QString("guide.%1").arg(guideId), guideObj);
+  if (!guide->m_composer) {
+    logger.warning() << "Composer failed";
+    return nullptr;
+  }
+
+  guard.dismiss();
+  return guide;
 }
 
 AddonGuide::AddonGuide(QObject* parent, const QString& manifestFileName,
                        const QString& id, const QString& name)
-    : Addon(parent, manifestFileName, id, name) {
+    : Addon(parent, manifestFileName, id, name, "guide") {
   MVPN_COUNT_CTOR(AddonGuide);
 }
 
-AddonGuide::~AddonGuide() {
-  MVPN_COUNT_DTOR(AddonGuide);
-
-  GuideModel::instance()->remove(id());
-}
+AddonGuide::~AddonGuide() { MVPN_COUNT_DTOR(AddonGuide); }
