--- src/models/guide.h
@@ -1,39 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#ifndef GUIDE_H
-#define GUIDE_H
-
-#include "composer.h"
-
-class QJsonObject;
-
-class Guide final : public QObject {
-  Q_OBJECT
-  Q_DISABLE_COPY_MOVE(Guide)
-
-  Q_PROPERTY(QString titleId MEMBER m_titleId CONSTANT)
-  Q_PROPERTY(QString subtitleId MEMBER m_subtitleId CONSTANT)
-  Q_PROPERTY(QString image MEMBER m_image CONSTANT)
-  Q_PROPERTY(Composer* composer MEMBER m_composer CONSTANT)
-
- public:
-  static Guide* create(QObject* parent, const QString& fileName);
-  static Guide* create(QObject* parent, const QJsonObject& obj);
-
-  const QString& titleId() const { return m_titleId; }
-
-  ~Guide();
-
- private:
-  explicit Guide(QObject* parent);
-
- private:
-  QString m_titleId;
-  QString m_subtitleId;
-  QString m_image;
-  Composer* m_composer = nullptr;
-};
-
-#endif  // GUIDE_H
