+++ src/models/tutorial.cpp
@@ -3,7 +3,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "tutorial.h"
-#include "guide.h"
 #include "leakdetector.h"
 #include "logger.h"
 #include "qmlengineholder.h"
