+++ src/ui/developerMenu/ViewFeatureList.qml
@@ -7,6 +7,7 @@ import QtQuick.Controls 2.14
 import QtQuick.Layouts 1.14
 
 import Mozilla.VPN 1.0
+import Mozilla.VPN.qmlcomponents 1.0
 import components 0.1
 
 import org.mozilla.Glean 0.30
