+++ tests/unit/testaddon.h
@@ -10,4 +10,7 @@ class TestAddon final : public TestHelper {
  private slots:
   void conditions_data();
   void conditions();
+
+  void guide_create_data();
+  void guide_create();
 };
