+++ tests/unit/unit.pro
@@ -76,8 +76,6 @@ HEADERS += \
     ../../src/models/feature.h \
     ../../src/models/featuremodel.h \
     ../../src/models/feedbackcategorymodel.h \
-    ../../src/models/guide.h \
-    ../../src/models/guidemodel.h \
     ../../src/models/helpmodel.h \
     ../../src/models/keys.h \
     ../../src/models/licensemodel.h \
@@ -142,7 +140,6 @@ HEADERS += \
     testandroidmigration.h \
     testcommandlineparser.h \
     testfeature.h \
-    testguide.h \
     testipaddress.h \
     testipaddresslookup.h \
     testipfinder.h \
@@ -197,8 +194,6 @@ SOURCES += \
     ../../src/models/feature.cpp \
     ../../src/models/featuremodel.cpp \
     ../../src/models/feedbackcategorymodel.cpp \
-    ../../src/models/guide.cpp \
-    ../../src/models/guidemodel.cpp \
     ../../src/models/helpmodel.cpp \
     ../../src/models/keys.cpp \
     ../../src/models/licensemodel.cpp \
@@ -262,7 +257,6 @@ SOURCES += \
     testandroidmigration.cpp \
     testcommandlineparser.cpp \
     testfeature.cpp \
-    testguide.cpp \
     testipaddress.cpp \
     testipaddresslookup.cpp \
     testipfinder.cpp \
